+++ quoftc.h
@@ -19,7 +19,7 @@
 #define NEW(type) ((type *) xmalloc(sizeof(type)))
 #define NEWC(type) ((type *) xcalloc(sizeof(type)))
 
-NORETURN PRINTF(2, 3) void fatal_error(uint16_t, const char *, ...);
+NORETURN PRINTF(2, 3) void fatal_error(unsigned, const char *, ...);
 NORETURN void internal_error(void);
 MALLOC void *xmalloc(size_t);
 MALLOC void *xcalloc(size_t);
